refactor(core,sql,py,wire): the client batch is the wire form, so the PK region is OPK and a payload column is its own region
`ZSetBatch` stored rows in a representation the wire did not carry, and
translated on every crossing: the PK column held native little-endian where
the region is OPK, and a STRING/BLOB column held owned cells where the region
holds 16-byte German strings against a blob arena. `ViewBuffers` existed to
paper over both, rebuilding a region list per view.

The client now stores what the wire carries. `PkColumn` holds OPK bytes with
private fields — routing hashes those bytes, so an un-encoded write would
mis-partition rather than error — and every append either encodes from native
values or copies OPK verbatim. `PkTuple` holds the same bytes, loses its
transport role to `gnitz_wire::control::{split_ctrl_key, join_ctrl_key}`, and
keeps one job: the row-identity key. `ZSetBatch.columns` is one region per
schema column against one `blob`; `ColData` and `ViewBuffers` are deleted.

Consequences: a WAL decode moves the PK region in one copy instead of a
per-row per-column transcode and no longer allocates a `Vec` per German cell;
an `ORDER BY` on a PK column is one `memcmp`; UPDATE/DELETE move a reply's
keys straight into the retraction batch. `german_string_cell_ok` moves to the
client decode, where the old materialization used to catch a bad heap extent.

Two guards were side effects of the old representation and are now written:
a string literal into a BLOB column stays rejected, and a STRING column still
extracts a `str` from Python where BLOB takes bytes. Both are tested.

Also: `Schema.pk_cols` becomes `Vec<u32>` (what every consumer takes) and
`pk_indices()` goes; `pk_index_single` returns `Option`; `ColumnDef`'s two
co-varying FK fields become `Option<FkTarget>`, which retires the `u64::MAX`
self-reference sentinel; `validate`'s NOT NULL sweep is an OR-fold; the
row-completeness check moves into `end_row`; and the `null_word_*` re-export
goes, leaving `gnitz_wire::` the one spelling.

+++ crates/gnitz-sql/src/hir/chain.rs
@@ -102,7 +102,7 @@ pub(crate) fn debug_assert_exchange_topology(circuit: &Circuit) {
 /// them to the DDL gateway — whose verdict is identical but names only a segment
 /// index, and which the ad-hoc fold's pre-map never reaches.
 pub(crate) fn schema_of(cols: &[ColumnDef], pk: PkArity, what: &str) -> Result<Arc<Schema>, GnitzSqlError> {
-    Schema::from_parts(cols.to_vec(), (0..pk).collect())
+    Schema::from_parts(cols.to_vec(), (0..pk as u32).collect())
         .map(Arc::new)
         .map_err(|e| GnitzSqlError::Unsupported(format!("{what}: {e}")))
 }
